Backport log rotation features from open source
This commit enables log rotation for `traffic.out` and `diags.log`.
This is a pretty massive internal change and as such it requires a
long explanation for what's been done and why. Refer to the
open source commits starting at 5cf0ae901502b7 (that's the open
source commit git hash, not YTS) to see more detailed explanations
and the related changes.