Backport TS-4165
TS-4165 Logging breaks if changing log format

TS wasn't calling `open_file()` on the LogFile it was trying
to rotate away. This was occurring when an existing log file's
formatting was changed in logs_xml.config and TS was restarted.
The result was TS not updating/creating the updated log file.