TS-4915: Crash from hostdb in PriorityQueueLess
This commit is based on the fix for TS-4938 which has yet to be merged.  Without that fix, I see the other crashes much more often.
I added an assert based on Bryan's ASAN build.  It showed that erase was being called with entry->index > _v.length().  By code
examination, it seems that the .pop() is harmful.  The previous erase call should have removed the relevant entry.  Running with this
fix now to see if it reduces the crash rate.

The changes in iocore/hostdb/P_RefCountCache.h and lib/ts/PriorityQueue.h are the relevant ones.