TS-5084 Make logcat follow file rotation
This patch makes `traffic_logcat -f` follow squid.blog even when
squid.blog is rotated. Previously, the tool did not do this at all.

This is implemented by keeping track of the _most_previous_ inode #
of squid.blog and comparing it with the most up to date value of
the inode # every now and then. If there is ever a difference, then
we close and reopen the log file.

Signed-off-by: Daniel Xu <dlxu2@yahoo.com>