Improve graceful shutdown.
Rather than sleeping in a signal handler, unify the signal handling
so that we always bounce the actual work up to the SignalContinuation.
Once we are there, then we can send a timed event to shutdown after the
timeot, or just cause a normal shutdown immediately.

On the exit path, stop the ProcessManager to keep the integration
tests happy.

(cherry picked from commit 693839a31e6e3839146f4d8aa80cd0a237a51db3)