YTSATS-1238: Build 156 is leaking memory
We added a call to trans->destroy() several places in Http1ClientSession to address
YTSATS-1126.  However destroy breaks the link from the state machine back to the ClientSession.
So the ClientSession will not be signaled to clean up on state machine shutdown.  Adjusted
destroy to remove the ClientSession's reference to the state machine, but not the
state machine's reference to the client session.