Simplify ProcessManager message and shutdown handling.
Simplify ProcessManager message handling to make it clearer when
we are stopping and wen we are dealing with a process manager error
pumping messages to traffic_manager. We do this by hoisting the
message reading cose into a helper function that just deals with
reading the data, and propagating the error up to the manager thread.
At that point, we can more easily know whether we are shutting down
or not.

(cherry picked from commit 6bcfa0888a996cd2d8523e422a0508c0ba2441a3)