Wake the process manager thread on shutdown.
If we are trying to stop the ProcessManager, the poll thread might
be sleeping, so we now send it SIGINT to break the sleep. This
causes it to notice that it should no longer be running earlier.

(cherry picked from commit 8c4a0eae65d9a49ef3017afa6e593417fb2be9a3)