Parent Selection -- Add option to mark down self specified as a parent.
Extends the proxy.config.http.parent_proxy.self_detect configuration
option so that a specified parent that has been detected as self may be
marked down instead of removed from the parent list.

This may be useful in some Parent Selection Consistent Hashing configurations
where we want the self-parent to remain in the hash ring but never be selected.
Simply removing the self-parent would change the distribution of parent
selection URLs across the remaining parents. This commit also makes this
the default behavior instead of filtering out self parents.