cdniuc is not a manditory claim
With Internet Draft 16 for uri signing, the cdniuc claim is not manditory. It
took the place of the manditory sub claim in draft 12, and the manditory nature
of the sub claim was still in effect. This change allows for tokens to not contain
the cdniuc claim and also renews the cdniuc and cdnistd claim on token renewal.