Fix FREELIST macros for AArch64
This revert the commit 8563409 and fix a bug of FREELIST_VERSION(_x).

The layout of FREELIST_POINTER:

-  0 ~ 47 bit : 48 bits, Virtual Address
                (1+47 bits for AMD64 and 48 bits for AArch64)
- 48 ~ 62 bit : 15 bits, Freelist Version
-      63 bit :  1 bits, The type of Virtual Address
                (0 = user space, 1 = kernel space)

The valid Virtual Address range for AMD64 is:

- 0000000000000000 ~ 00007fffffffffff for user space
- ffff800000000000 ~ ffffffffffffffff for kernel space

The 47th bit can be used to indicate the type of VA.

And the one (64k page size) of the valid VA range for AArch64 is:

- 0000000000000000 ~ 0000ffffffffffff for user space
- ffff000000000000 ~ ffffffffffffffff for kernel space

We can see that the actual length of the Virtual Address is 48 bits on
AArch64 and 47 bits on AMD64. Therefore, we cannot use the 47th bit to
indicate the type of Virtual Address, which requires an extra bit.

The highest bit would be a good choice because we have done that before.