Removes the abort() from header_rewrite, and try to deal with errors
Bryan suggested using exceptions here, rather than just producing errors
in the log files. As ugly as it is, it does seem to work ... But I blame
him for all future issues! Long term, I think it'd be better to eliminate
the exceptions here, and fix all the various layers of returning and
perculating errors all the way back up through the stack.