TS API - Adds the TSHttpTxnNextHopNameGet() function.
The new API function TSHttpTxnNextHopNameGet() is added to the API.
It is also added to the Lua plugin API as
ts.server_request.server_addr.get_nexthop_name(). This new call
operates in the same context as TSHttpTxnNextHopAddrGet() except
that it returns the host name string set by parent selection
for the next hop host rather than the socket (which provides IP
address, port, and address family).

The use case for this is when you must modify the send request
header based on the selected parent *and* mapping based on a
numeric IP is not feasible due to dynamic address changes
or a large number of addresses matching a parent host.