cachekey: added --key-type (for parent selection)
Added ability to apply all transformations, available for modifying
the cache key, to parent selection URL:
* --key-type=cache_key - apply transformations to  cache key
* --key-type=parent_selection_url - apply transformations to
  parent selection URL

TODO/TBD: After this change all transformations can be applied not
only to the cache key but to parent selection URL as well. It would
make sense to give the cachekey plugin a new, more suitable name.