Removing traffic_cop reference in Admin Guide introduction.
The introduction to our Administrator's Guide referenced three processes but
only described two: traffic_manager and traffic_server. The third process was
probably traffic_cop. But we currently now only run with traffic_server and
traffic_manager. This adjusts the number of referenced processes from three to
two and removes the image which is not as helpful anymore with only two
processes that collaborate.