Reverse debug specific check.
debug_flag && diags->on() is almost an atomic operation,
while is_debug_tag_set requires Diags to run a regular expression.

When debug_flag is set to 1, we always want to print the debug log,
regardless of whether the tag is set or not, which makes the
is_debug_tag_set check useless, and expensive.

This change flips the conditional to check if debug_flag is 1 first,
skipping the tag check if it is.

Signed-off-by: David Calavera <david.calavera@gmail.com>