auto delete rolled log file fixes
This fixes auto delete registration for:

- core log files that were not getting deleted (such as error.log and
  manager.log) as well as...
- plugin log files.

To make log deletion registration more automatic, I placed it in LogObject
initialization so that it doesn't have to happen in as many disparate places.
By doing this, plugin log auto delete comes for free.  This is why nothing in
the core plugin implementation needed to be changed to register their log
files, for instance. Not all log objects are managed by LogObject, however, so
we still have some manual calls to registration in LogConfig for core log files
(such as traffic.out and diags.log, for example).

Before this change, LogConfig manipulated deleting_info to create and maintain
candidates. To encapsulate this logic, I created RolledLogDeleter and changed
LogConfig to interact with an instance of that. This simplified the LogConfig
logic while also enabling unit testability of the main log deletion feature.

This also:

- Adds a rolling_max_count autest.
- Fixes a shutdown leak concerning LogDeletingInfo.
- Fixes a paramater mismatch between TextLogObject and LogObject.