SSL: Introduce proxy.config.ssl.server.prioritize_chacha
OpenSSL 1.1.1 introduces a new server side option (SSL_OP_PRIORITIZE_CHACHA)
that temporarily reprioritize ChaCha20-Poly1305 ciphers to the top of
the server cipher list if a ChaCha20-Poly1305 cipher is at the top of
the client cipher list. This helps those clients (e.g. mobile) use
ChaCha20-Poly1305 if that cipher is anywhere in the server cipher list;
but still allows other clients to use AES and other ciphers