YTSATS-3761: Fix crash in Lua host lookup
I think I miss reading the documentation - https://github.com/apache/trafficserver/blob/master/doc/developer-guide/api/functions/TSHostLookup.en.rst

I was not aware of the case when the call can be synchronous.

Also we need to yield a luaVM when we need to wait for an answer. And we should not yield if the call is synchronous.
And then in the handler, we need to see if we have yielded or not. So we can resume only if we have yielded.