Fix ConfigManager memory leaks. (#536)
This fixes an ASan reported leak of ConfigManager::configName.  It used
to be strdup'd but not freed in the destructor. This simply changes it
to a std::string.

ASan also reported a leak in AddConfigFilesHere which is fixed with an
ats_free as well.

This also fixes the corresponding leaks in our mgmt2 code as well.